+++ public/index.html
@@ -31,7 +31,7 @@
     <div id="root"></div>
     <!--
       This HTML file is a template.
-      If you open it directly in the browser, you will see an empty page.
+      If you open it directly in the browser, you will see an empty pages.
 
       You can add webfonts, meta tags, or analytics to this file.
       The build step will place the bundled scripts into the <body> tag.
