--- src/Api/HomeData/Data.js
@@ -1,26 +1,0 @@
-
-
-const Trendingimg =[
-
-    {
-        Trending:'https://www.zipcodewilmington.com/hs-fs/hubfs/SI-Images/landing-pages/interface-icons-people-162px.png?width=162&name=interface-icons-people-162px.png'
-    },
-    {
-        Trending:'https://www.zipcodewilmington.com/hs-fs/hubfs/SI-Images/landing-pages/interface-icons-people-162px.png?width=162&name=interface-icons-people-162px.png'
-    },
-    {
-        Trending:'https://www.zipcodewilmington.com/hs-fs/hubfs/SI-Images/landing-pages/interface-icons-people-162px.png?width=162&name=interface-icons-people-162px.png'
-    }
-];
-
-
-
-
-
-const Data={
-    Trendingimg
-};
-export default  Data;
-
-
-
