--- src/Assets/Home/Icons.js
@@ -1,1 +1,0 @@
-export const homepc=require('./homepc.png');
