+++ src/api/Data.js
@@ -1,0 +1,36 @@
+import Image from "../pages/home/assets";
+
+
+const trending =[
+
+    {
+        title:'1',
+        url:Image.movie1
+    },
+    {
+        title:'1',
+        url:Image.movie1
+    },
+    {
+        title:'1',
+        url:Image.movie1
+    },
+    {
+        title:'1',
+        url:Image.movie1
+    },
+
+
+];
+
+
+
+
+
+const Data={
+    trending
+};
+export default  Data;
+
+
+
