+++ src/assets/index.js
@@ -1,0 +1,11 @@
+import banner from './/homepc.png';
+import car from './/car.png';
+import movie1 from './/whachlist1.png';
+
+const Image = {
+    banner,
+    car,
+    movie1
+};
+
+export default Image;
