+++ src/components/index.js
@@ -1,0 +1,7 @@
+import Trending from './Trending';
+import RecentPlaying from './RecentPlaying';
+
+export {
+    Trending,
+    RecentPlaying
+}
