+++ src/index.js
@@ -4,7 +4,7 @@ import {BrowserRouter as Router} from 'react-router-dom';
 import './index.css';
 import App from './App';
 import * as serviceWorker from './serviceWorker';
-import Routes from '../src/Routes/routes'
+import Routes from './routes/routes'
 
 
 ReactDOM.render(
