+++ src/pages/home/components/RecentPlaying.js
@@ -1,0 +1,19 @@
+import React from "react";
+
+
+const RecentPlaying=()=>{
+    return(
+
+        <div className={""}>
+            <div>
+
+            </div>
+
+            <div>
+
+            </div>
+
+        </div>
+    );
+};
+export default RecentPlaying;
