+++ src/pages/test/Test.css
@@ -1,0 +1,12 @@
+/*.img{*/
+
+    /*justify-content: center;*/
+    /*align-items: center;*/
+/*}*/
+
+
+/*.Logo{*/
+    /*margin-right:700px;*/
+    /*margin-top:-330px;*/
+    /*color: white;*/
+/*}*/
