+++ src/pages/test/Test.js
@@ -1,0 +1,20 @@
+// import React from "react";
+// import {useState} from "react";
+// import Image from "./assets";
+// import {Layout, Menu, Breadcrumb} from 'antd';
+// import {Row, Col, Divider} from 'antd';
+//
+// const {Header, Content, Footer} = Layout;
+//
+//
+// const Test = (props) => {
+//
+//     return (
+//         <Layout>
+//
+//         </Layout>
+//
+//
+//     );
+// };
+// export default Test;
