+++ src/pages/test/assets/index.js
@@ -1,0 +1,7 @@
+// import beauty from './beauty.png';
+//
+// const Image = {
+//     beauty,
+// };
+//
+// export default Image;
