+++ src/pages/test/components/index.js
@@ -1,0 +1,5 @@
+// import Trending from './Trending';
+//
+// export {
+//     Trending,
+// }
