+++ src/routes/routes.js
@@ -2,7 +2,7 @@ import React, { Component } from "react";
 import { Router, Switch, Route } from "react-router-dom";
 import App from "../App";
 import history from './history';
-import Home from "../Page/Home"
+import Home from "../pages/home/Home"
 export default class Routes extends Component {
     render() {
         return (
